i think I got it

+++ src/main/kotlin/TempSchedule.kt
@@ -8,12 +8,12 @@ fun weightedCoinFlip(trueProbability: Double) =
 
 
 
-class TempSchedule(val maxTemp: Int, val temperatureSequence: IntArray)  {
+class TempSchedule(val maxTemp: Int, val temperatureSequence: DoubleArray)  {
 
     private var index = -1
 
     val heat get() = temperatureSequence[index]
-    val ratio get() = temperatureSequence[index].toDouble() / maxTemp.toDouble()
+    val ratio get() = temperatureSequence[index]  / maxTemp.toDouble()
 
     fun next(): Boolean {
         if (index == (temperatureSequence.size-1)) return false
